Rename tlsenum/cipher_suites.py to tlsenum/mappings.py.
